Add philospohers lock fixes

+++ philo/src/common/time.c
@@ -6,18 +6,18 @@
 /*   By: meskelin <meskelin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/06 17:50:33 by meskelin          #+#    #+#             */
-/*   Updated: 2023/06/13 16:44:58 by meskelin         ###   ########.fr       */
+/*   Updated: 2023/06/19 17:12:31 by meskelin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../headers/philo.h"
 
-size_t	time_ms(void)
+unsigned long long	time_ms(void)
 {
 	struct timeval			tv;
 
 	if (gettimeofday(&tv, NULL) == 0)
-		return ((tv.tv_sec) * (unsigned int)1000 + (tv.tv_usec) / 1000);
+		return (((tv.tv_sec) * (unsigned int)1000) + ((tv.tv_usec) / 1000));
 	else
 		return (0);
 }
